+++ app/scripts/controllers/transactions/index.js
@@ -32,7 +32,7 @@ import {
 
 abiDecoder.addABI(abi)
 
-const SIMPLE_GAS_COST = '0x5208' // Hex for 21000, cost of a simple send.
+const SIMPLE_GAS_COST = '0xC738' // Hex for 51000, cost of a simple send.
 const MAX_MEMSTORE_TX_LIST_SIZE = 100 // Number of transactions (by unique nonces) to keep in memory
 
 /**
@@ -349,7 +349,7 @@ export default class TransactionController extends EventEmitter {
         from,
         to: from,
         nonce,
-        gas: '0x5208',
+        gas: '0xC738',
         value: '0x0',
         gasPrice: newGasPrice,
       },