+++ ui/app/components/app/transaction-list-item-details/tests/transaction-list-item-details.component.test.js
@@ -15,7 +15,7 @@ describe('TransactionListItemDetails Component', function () {
       status: 'confirmed',
       txParams: {
         from: '0x1',
-        gas: '0x5208',
+        gas: '0xC738',
         gasPrice: '0x3b9aca00',
         nonce: '0xa4',
         to: '0x2',
@@ -56,7 +56,7 @@ describe('TransactionListItemDetails Component', function () {
       status: 'confirmed',
       txParams: {
         from: '0x1',
-        gas: '0x5208',
+        gas: '0xC738',
         gasPrice: '0x3b9aca00',
         nonce: '0xa4',
         to: '0x2',
@@ -99,7 +99,7 @@ describe('TransactionListItemDetails Component', function () {
       status: 'confirmed',
       txParams: {
         from: '0x1',
-        gas: '0x5208',
+        gas: '0xC738',
         gasPrice: '0x3b9aca00',
         nonce: '0xa4',
         to: '0x2',
@@ -141,7 +141,7 @@ describe('TransactionListItemDetails Component', function () {
       hash: '0xaa',
       txParams: {
         from: '0x1',
-        gas: '0x5208',
+        gas: '0xC738',
         gasPrice: '0x3b9aca00',
         nonce: '0xa4',
         to: '0x2',