+++ ui/app/pages/confirm-transaction-base/confirm-transaction-base.component.js
@@ -177,7 +177,7 @@ export default class ConfirmTransactionBase extends Component {
       }
     }
 
-    if (hexToDecimal(customGas.gasLimit) < 21000) {
+    if (hexToDecimal(customGas.gasLimit) < 51000) {
       return {
         valid: false,
         errorKey: GAS_LIMIT_TOO_LOW_ERROR_KEY,